Prevent unnecessary duplicate byte arrays
Creating an Image from a byte array previously would create a stream
from the byte array, and then another byte array from that stream, from
which other streams would be created.

Flip the relationship between the methods that expect byte arrays and
streams, so that when a byte array is given, streams are created
directly from it rather than creating an additional duplicate byte
array.