Re-orient image based on Exif orientation data in IFD0 only
IFD0 is the only Exif metadata directory where the rotation tag for the
actual image may be located.

IFD1 is the metadata directory for the thumbnail image. It may contain
the same orientation tag as in IFD0, but with different value.

This change makes Scrimage to re-orient the image based on orientation
metadata in IFD0 directory only.

For more about IFD0 and IFD1, Exiv2 tool has a nice summary:
http://www.exiv2.org/tags.html